# openapi.yaml
openapi: 3.0.0
info:
  title: Feedback 360 Survey API
  version: 2.0.0
  description: |
    This API specification contains information around the Feedback 360 Survey
    API developed by the App Dev Practice.

    The Feedback 360 Survey API enables the management of feedback surveys for
    customer engagements. A feedback survey is N/A to 5 point scale of competency around
    the technologies used for a project. Each added employee will be able to enhance the
    survey by adding technologies, softskills, and freetext for each of his or her
    colleagues. The will enable Red Hat to better understand multiple aspects of
    a consultant's skillset.
  contact:
    email: appdevpractice@redhat.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
servers:
  - url: "http://localhost:3007"
    description: LOCAL
  - url: "http://survey-service-consultant-360-dev.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on DEV Cluster
  - url: "http://survey-service-consultant-360-prod.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on PROD Cluster
paths:
...
  "/surveySubmission/{surveySubmissionId}":
    get:
      tags:
        - surveySubmission
      parameters:
        - name: surveySubmissionId
          description: ID of the surveyGroup as a GUID.
          schema:
            type: string
            format: uuid
            example: f9238beb-9649-4983-9059-4f0ee372d56e
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveySubmissionWithSkillRating"
          description: Survey submission associated with the survey submission resource ID
        "404":
          description: An survey submission matching the given submission ID was not found.
      operationId: getSurveySubmissionById
      summary: Fetch the survey submission associated with the survey submission resource ID
      description: Returns a survey submission matching the submission ID.
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurveySubmissionWithSkillRating"
      tags:
        - surveySubmission
      parameters:
        - name: surveySubmissionId
          description: ID of the employee assignment as a GUID.
          schema:
            type: string
            format: uuid
            example: f9238beb-9649-4983-9059-4f0ee372d56e
          in: path
          required: true
      responses:
        "200":
          description: >-
            The survey submission assignment represented by the supplied GUID has been updated
            with the contents of this request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveySubmissionEntity"
        "400":
          description: >-
            The survey submission was not updated. Something is wrong with the request like a
            missing required field.
      operationId: updateSurveySubmissionByID
      summary: |
        Update a survey Submission in Feedback 360 Survey API.
      description: >
        This representation of the survey submission contained in this request will replace  the
        survey submission located at this URI. This operation is idempotent. TODO
    delete:
      tags:
        - surveySubmission
      parameters:
        - name: surveySubmissionId
          description: ID of the survey submission as a GUID.
          schema:
            type: string
            format: uuid
            example: f9238beb-9649-4983-9059-4f0ee372d56e
          in: path
          required: true
      responses:
        "204":
          description: The survey submission represented by the supplied GUID is disabled.
      operationId: disableSurveySubmission
      summary: Disable a surveySubmission from Feedback 360 Survey API.
      description: >
        This operation will disable a survey submission and all child resources  associated with it.
        This operation is idempotent.
  "/surveySubmission/reviewedBy/{employeeId}":
    get:
      tags:
        - surveySubmission
      parameters:
        - name: employeeId
          description: Employee's RedHat Kerberos ID.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveySubmission"
          description: Survey submissions associated with the survey submission resource ID
        "404":
          description: Survey submissions matching the given employee ID were not found.
      operationId: getReviewedByByEmployeeId
      summary: Fetch the survey submissions reviewed by the given employee Id.
      description: Returns all survey submissions that have been reviewed by the employee ID.
  "/surveySubmission/reviewedOn/{employeeId}":
    get:
      tags:
        - surveySubmission
      parameters:
        - name: employeeId
          description: Employee's RedHat Kerberos ID.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveySubmission"
          description: Survey submissions associated that the employee was reviewed on.
        "404":
          description: Survey submissions matching the given employee ID were not found.
      operationId: getReviewedOnByEmployeeId
      summary: Fetch the survey submissions that the given employee Id was reviewed on.
      description: Returns all survey submissions that the employee has been reviewed on.
  "/surveySubmission/surveyGroup/{surveyGroupId}":
    get:
      tags:
        - surveySubmission
      parameters:
        - name: surveyGroupId
          description: ID of the surveyGroup as a GUID.
          schema:
            type: string
            format: uuid
            example: f9238beb-9649-4983-9059-4f0ee372d56e
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveySubmission"
          description: Survey submissions associated with the Survey Group resource ID
        "404":
          description: Survey submissions matching the given Survey Group ID were not found.
      operationId: getSurveySubmissionsBySurveyGroupId
      summary: Fetch the survey submissions associated with the Survey Group.
      description: Returns all survey submissions that are part of the SurveyGroup.
...
components:
  schemas:
    EmployeeEntity:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - |-
          @javax.persistence.Table(
            name = "employees",
            uniqueConstraints = {
              @javax.persistence.UniqueConstraint(columnNames = {"email"})
            }
          )
      allOf:
        - $ref: "#/components/schemas/Employee"
      properties:
        assignments:
          description: Projects for which the employee has been assigned to a SurveyGroup
          items:
            $ref: "#/components/schemas/EmployeeAssignment"
          type: array
          x-java-field-annotations:
            - '@javax.persistence.OneToMany(mappedBy = "employee")'
...
    SurveyGroupEntity:
      required:
        - opportunityId
        - projectName
        - projectCreatorId
        - tsmId
      type: object
      allOf:
        - $ref: "#/components/schemas/SurveyGroup"
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - '@javax.persistence.Table(name = "survey_groups")'
      properties:
        assignments:
          type: array
          x-java-field-annotations:
            - "@javax.persistence.OneToMany"
            - '@javax.persistence.JoinColumn(name = "survey_group_id")'
          items:
            $ref: "#/components/schemas/EmployeeAssignment"
        disabled:
          description: A flag indicating if this Survey Group is disabled
          type: boolean
          x-java-field-annotations:
            - "@javax.persistence.Column(nullable = false)"
          default: false
          example: false
    SurveyGroup:
      required:
        - opportunityId
        - projectName
        - projectCreatorId
        - tsmId
      type: object
      x-java-class-annotations:
        - "@javax.persistence.MappedSuperclass"
        - "@javax.persistence.EntityListeners(org.springframework.data.jpa.domain.support.AuditingEntityListener.class)"
      properties:
        id:
          description: A GUID that uniquely identifies the SurveyGroup.
          type: string
          format: uuid
          example: f1ad7649-eb70-4499-9c82-a63fe2c6dc71
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID", strategy =
              "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable =
              false)'
        createdDate:
          description: The date the project was created in the Feedback 360 Survey API in UTC.
          type: string
          format: date-time
          readOnly: true
          example: "2020-01-29"
          x-java-field-annotations:
            - "@org.springframework.data.annotation.CreatedDate"
        createdBy:
          description: The employee id (Kerberos) of the user that created the project.
          type: string
          example: janedoe
          x-java-field-annotations:
            - "@org.springframework.data.annotation.CreatedBy"
        modifiedDate:
          description: The date the project was last modified in the Feedback 360 Survey API in UTC.
          type: string
          format: date-time
          readOnly: true
          example: "2020-01-29"
          x-java-field-annotations:
            - "@org.springframework.data.annotation.LastModifiedDate"
        modifiedBy:
          description: The employee id (Kerberos) of the user that last modifed the project.
          type: string
          example: janedoe
          x-java-field-annotations:
            - "@org.springframework.data.annotation.LastModifiedBy"
        opportunityId:
          description: The ID of the Opportunity from PSA.
          type: string
          example: 3456NAS
          x-java-field-annotations:
            - "@javax.persistence.Column(unique = true)"
        projectName:
          description: The name of the project.
          type: string
          example: NASA App Modernization
        projectCreatorId:
          description: The email address of the creator of the project.
          type: string
          example: janedoe@redhat.com
        tsmId:
          description: The email address of the TSM in charge of the project.
          type: string
          example: timmytsm@redhat.com
        defaultSkills:
          type: array
          x-java-field-annotations:
            - "@javax.persistence.ManyToMany"
            - |-
              @javax.persistence.JoinTable(
                  name = "default_survey_skills",
                  joinColumns = @javax.persistence.JoinColumn(name = "survey_group_id"),
                  inverseJoinColumns = @javax.persistence.JoinColumn(name = "skill_id"),
                  uniqueConstraints = @javax.persistence.UniqueConstraint(columnNames = {"survey_group_id", "skill_id"})
              )
          items:
            $ref: "#/components/schemas/Skill"
...
    SurveySubmissionEntity:
      type: object
      allOf:
        - $ref: "#/components/schemas/SurveySubmission"
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - '@javax.persistence.Table(name = "survey_submissions",
          uniqueConstraints={
          @javax.persistence.UniqueConstraint(columnNames={"survey_group_id", "survey_author_id", "survey_subject_id"})
          })'
      properties:
        surveyGroup:
          description: A GUID that uniquely identifies the project.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.ManyToOne"
            - '@javax.persistence.JoinColumn(name = "survey_group_id")'
          allOf:
            - $ref: "#/components/schemas/SurveyGroupEntity"
        ratings:
          description: SkillRatings for the subject employee
          type: array
          x-java-field-annotations:
            - "@javax.persistence.OneToMany(cascade = javax.persistence.CascadeType.ALL, orphanRemoval=true)"
            - '@javax.persistence.JoinColumn(name = "survey_submission_id")'
          items:
            $ref: "#/components/schemas/SkillRatingEntity"
        surveyAuthorEntity:
          description: The Kerberos Id for the employee that is answering the questions of the survey.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.ManyToOne"
            - '@javax.persistence.JoinColumn(name = "survey_author_id")'
          allOf:
            - $ref: "#/components/schemas/EmployeeEntity"
        surveySubjectEntity:
          description: A Kerberos Id for the employee that is being reviewed.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.ManyToOne"
            - '@javax.persistence.JoinColumn(name = "survey_subject_id")'
          allOf:
            - $ref: "#/components/schemas/EmployeeAssignment"
        disabled:
          description: A flag indicating if this Survey Submission is disabled
          type: boolean
          x-java-field-annotations:
            - "@javax.persistence.Column(nullable = false)"
          default: false
          example: false
    SurveySubmissionWithSkillRating:
      type: object
      allOf:
        - $ref: "#/components/schemas/SurveySubmission"
      properties:
        surveyGroup:
          description: A GUID that uniquely identifies the project.
          type: object
          allOf:
            - $ref: "#/components/schemas/SurveyGroup"
        ratings:
          description: SkillRatings for the subject employee
          type: array
          items:
            $ref: "#/components/schemas/SkillRating"
    SurveySubmission:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.MappedSuperclass"
      properties:
        id:
          description: A GUID that uniquely identifies a survey submitted for a project.
          type: string
          format: uuid
          example: f1ad7649-eb70-4499-9c82-a63fe2c6dc71
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID",strategy = "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable = false)'
        surveyAuthor:
          description: Employee that is reviewing.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.Transient"
          allOf:
            - $ref: "#/components/schemas/Employee"
        surveySubject:
          description: Employee that is being reviewed.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.Transient"
          allOf:
            - $ref: "#/components/schemas/EmployeeAssignment"
        submissionDate:
          description: The date the survey was submitted to the Feedback 360 Survey API in UTC.
          type: string
          format: date-time
          example: "2020-02-29"
        submissionStatus:
          description: Returns a status of "Complete" or "Incomplete"
          type: string
          example: Complete
          enum:
            - Complete
            - Incomplete
        feedback:
          description: Any feedback the author of the survey would like to include in the rating.
          type: string
          example: John was a pleasure to work with and extremely knowledgeable in REST.
    SkillRatingEntity:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - '@javax.persistence.Table(
          name = "skill_rating",
          uniqueConstraints={
          @javax.persistence.UniqueConstraint(columnNames={"skill_id", "survey_submission_id"})
          })'
      allOf:
        - $ref: "#/components/schemas/SkillRating"
    SkillRating:
      required:
        - skillId
      type: object
      x-java-class-annotations:
        - "@javax.persistence.MappedSuperclass"
      properties:
        id:
          description: A GUID that uniquely identifies a skill rathing.
          type: string
          format: uuid
          example: f1ad7649-eb70-4499-9c82-a63fe2c6dc71
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID",strategy = "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable = false)'
        skill:
          description: A GUID that uniquely identifies a skill.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.ManyToOne"
            - '@javax.persistence.JoinColumn(name = "skill_id", nullable=false)'
          allOf:
            - $ref: "#/components/schemas/Skill"
        rating:
          description: >-
            The rating -1 -> 4 of the skill. -1 - 'Not Applicable', 0 - No Experience, 1 - Foundational, 2 - Experienced, 3 - Advanced, 4 - Expert
          type: integer
          minimum: -1
          maximum: 4
...
    EmployeeAssignment:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - '@javax.persistence.Table(
          name = "employee_assignments",
          uniqueConstraints={
          @javax.persistence.UniqueConstraint(columnNames={"employee_id", "survey_group_id"})
          })'
      properties:
        id:
          description: A GUID that uniquely identifies the project.
          type: string
          format: uuid
          example: f9238beb-9649-4983-9059-4f0ee372d56e
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable = false)'
        employee:
          allOf:
            - $ref: "#/components/schemas/EmployeeEntity"
          type: object
          x-java-field-annotations:
            - "@javax.persistence.ManyToOne(cascade = javax.persistence.CascadeType.REMOVE)"
            - '@javax.persistence.JoinColumn(name = "employee_id", nullable=false)'
        surveyGroup:
          description: The unique ID of the SurveyGroup associated with the opportunity.
          type: object
          x-java-field-annotations:
            - "@javax.persistence.ManyToOne"
            - '@javax.persistence.JoinColumn(name = "survey_group_id", nullable=false)'
          allOf:
            - $ref: "#/components/schemas/SurveyGroupEntity"
        startDate:
          description: The date the employee started the project.
          type: string
          format: date-time
          example: "2020-01-29"
        endDate:
          description: The employee's end date on the project.
          type: string
          format: date-time
          example: "2020-01-29"
        billableRole:
          description: Role that the employee is billed for on the project..
          type: string
          example: Consultant
...